Feature: RD-293 - Proposal linking card

+++ features/proposals/providers/multisig/utils.ts
@@ -66,7 +66,7 @@ const getProposalCreationData = async function (
   if (!logs?.length) throw new Error("No creation logs");
 
   const log = logs[0];
-  const block = log.blockNumber as bigint;
+  const block = log.blockNumber;
   const tx = log.transactionHash;
 
   const logData: ProposalCreatedLogResponse = log.args as ProposalCreatedLogResponse;
